refactor(db): relocate service modules to appdb directory
- Move characterCardService, presetService, and worldBookService to src/database/appdb/
- Update all component and composable imports to reflect the new directory structure
- Simplify sidebar configuration event listener logic in App.vue
- Update dynamic import paths in database utility functions to maintain registry functionality
