+++ src/composables/worldbook/useWorldBookCollection.ts
@@ -3,7 +3,7 @@ import { ElMessage, ElMessageBox } from 'element-plus';
 import type { WorldBook, WorldBookCollection, WorldBookEntry } from '@/types/worldbook';
 import { v4 as uuidv4 } from 'uuid';
 import { processImportedWorldBookData } from './entry/useWorldBookEntryData';
-import { worldBookService } from '../../database/worldBookService';
+import { worldBookService } from '../../database/appdb/worldBookService';
 import type { StoredWorldBook } from '../../database/db';
 import { nowIso, formatDateTime } from '@/utils/datetime';
